Introduce new style for help pages, compatible with the new style of the overview pages.
* src/html/classic/img/help.svg: New.

* CMakeLists.txt: Install the new help icon.

* src/html/classic/css/gsa-style.css: Add definitions for h1 and h2 of
section-header where so far only h3 was defined. Also add an (so far)
empty section-box to be used in a box below a section header.
Finally added small-icon for 16x16 version of SVG icons.

* src/html/classic/help.xsl (help-header): New macro for the head of help pages.
Changed all of the pages to use the new header. Some links like for "new" dialogs
were removed as they don't make sense with the new gui.



git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@23760 423fd1db-d629-0410-8442-d21db03e70f4