Turn Filter section into a toolbar-like compact element.
* src/html/classic/css/gsa-style.css: Slight vertical adjustment of
icon toolbar.

* src/html/classic/omp.xsl (filter-window-part): Drop the reload icon because
the drop down box will always execute the reload automatically. Instead place
here the expand icon for the filter details dialog. And prepend "Filter:" to
the entire line.
(list-window): Drop the entire section header for "filter" and shift the
the filter up to the same level as the icon toolbar.



git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@23792 423fd1db-d629-0410-8442-d21db03e70f4