Fix help pages icons
* src/html/classic/img/secinfo.svg: Rename secinfo.svg to allinfo.svg. The
info type for all sec info is allinfo. The info type is used at several places

to create the links and icon locations for the help page.
* src/html/classic/graphics.xsl: Use allinfo.svg instead of secinfo.svg.

* src/html/classic/help.xsl: Update types for reports, results, oval defs and
cert advisories. This fixes displaying the correct icon at the help pages.

* src/html/classic/omp.xsl: Use allinfo as info type for all sec info. This
fixes the link to the help page of all sec info.

* CMakeLists.txt: Update installation of renamed secinfo.svg  file.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@24467 423fd1db-d629-0410-8442-d21db03e70f4