Refactoring of http request handling and code splitting
* src/gsad.c, src/gsad_http_handler.c, src/gsad_http_handler.h: Move http
request handling from gsad.c to gsad_http.c. Remove a lot of duplicate code.
Introduce a http_handler struct that acts like a chain of functions. Each
function in the chain can be called to handle the current request. In the
function it can be decided to create a response or leave the response for the
next http_handler in the chain. Additionally it is possible to only pass some
data to the following http_handlers in the chain.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@26804 423fd1db-d629-0410-8442-d21db03e70f4