Fix parsing user_tags of info models
The response of a info entity contains a sub element with the actual
data. This includes the user_tags. Therefore to parse the user tags
correctly in the Model base class the actual data must be moved to the
main object BEFORE parsing the properties in the Model calss.

This adds a generic data movement into the Info base class and calls
Model.parseProperties afterwards.