Add a State component
The new State component uses the render props pattern to store a state.
The state can be changed by calling onValueChange(value, name).
Afterwards value will be stored in the passed state e.g. state[name] ===
value.

Example for using the State component:

<State>
{({
  state,
  onValueChange,
}) => {
  const {myvalue} = state;
  return (
    <TextField
      value={myvalue}
      name="myvalue"
      onChange={onValueChange}
    />
  );
}}
</State>

With the state component it will not be necessary to re-implement the
state handling in components using forms.