Add new permission loader function for roles, groups and users
Two different permission loaders exists. A loader for roles, groups and
users and one for the others. Roles, groups and users can be connected
to the permission via subject too. All others only via the resource.

In future all components using the EntityContainer should set the
desired loaders and the loadPermissions property should be removed.