Add margin-left to CheckBoxRadio
Dividers don't work properly in dialogs, as the margin for the radio
buttons is calculated from a slight overlap with the previous component.
By adding a left margin, the overlap is fixed (display-wise, not for
calculation of position) and insertion of a lot of (not-properly
working) Dividers can be avoided.