Fix resizing in firefox
Firefox has a different behavior when calulating the height of the
scrollable content flexbox automatically compared to chrome. Therefore
resizing the dialogs caused rendering the content with full height.
Setting overflow to hidden fixes this behavior and the height is
adjusted to the max. possible height for the flexbox content and shows
the scrollbars instead.