Reduce gsad_message_new to minimum needed.
Now that only a single call of gsad_message_new happens,
all parts of that function can be removed which are not
used by this call (credential, function, line, xml_flag).

Also remove the proto from the header file which is not
needed anymore and due to this move up the function before
first use.