Update Filter equals method
Compare filters with multiple terms with same keyword correctly. The new
algorithm can increase the comparision time heavily because it has to
iterate over all terms several times.

To improve this situation in future we should parse and, or, ...
differently. They should be parsed into a special FilterTerm object that
connects to terms.