Add storybook for component development
"Storybook is a development environment for UI components. It allows you
to browse a component library, view the different states of each
component, and interactively develop and test components." (quote from
https://github.com/storybooks/storybook)

It can be used by running yarn storybook. This command will start a http
server running on localhost and port 9009.

Currently there are only stories for the new Select component.