Fix setting the current filter of the Powerfilter component
If the new filter prop is undefined the filter is reset and the state
must be updated. If no current filter is set and a new filter comes in
or the current filter id is different then the new filter id we must
also update the state and display the new filter.