Get rid of react warning
Get rid of the "Stateless function components cannot be given refs"
warning. The Dashboard is a composed component which root component is a
function component. Therefore it's not possible to use the ref here. To
avoid this all used HOCs of Dashboard would have to implement the
innerRef pattern to be able to get the ref to the actual Dashboard
component.

But here we can avoid that because calling the reload method of
Dashboard wasn't used at all. Therefore we can remove the ref and
convert the DashboardWrapper also to a stateless function component.