Update default behaviour of Select component
Raise warning if items are set and value is not in items. If items are
loaded async and therefore are not available for the initial render of a
Select component items prop MUST be set to undefined. undefined now
means there are no items yet. An empty items array means no items are
available.