Refactor Grid component
Version #2 of the Grid component.

Extract styled dropables as row component and the styled dragables as
items component to own js modules.
Add functions to create rows and items to abstract the expected row and
item objects. This also allows to assign unique ids for rows and items
on creation via uuid module.

Update the sorting. If an item is moved from one row to another put it
into the destination index instead of the end.

Always add a dropable empty row at the bottom of the list.

Remove the last row if it doesn't contain any item.