Add loopback addresses in add_local_addresses.
The default loopback addresses localhost, 127.0.0.1 and ::1 are now
added in add_local_addresses, which now also has an option not to add
the other interface addresses.
Also, add_local_addresses is now a separate step from gsad_address_init,
which only adds the explicit host addresses to the ones allowed in
the Host header.