Return gsad_message if password is invalid
The "Edit My Settings" does not need the settings included in the
response when submitting the changes, so it will only return the error
message if the old password is invalid.
The HTTP return code has also been changed to 400 as it makes more sense
for an invalid user input than 500.