Update SaveDialog to support controlled and uncontrolled values
Setting defaultValues (replaces initialData prop which is kept for
backwards compability) allows to the initial values for the state of the
SaveDialog. Not putting the defaultValues data also into values prop
means these defaultValues are used uncontrolled. Setting something in
values means the data is controlled by the parent component and is not
put in the state of the dialog.