Don't auto select a scp credential id
In most cases the user doesn't want to select this credential id for scp
method. Also it may be dangerous to pre-select one because a use may not
expect this behavior. Additionally the list of credentials may contain
not valid credential types for scp method. Thus the initial selected
credential may be wrong for scp method.