Reset task state variables when opening the create task dialog
It may be possible that the user did edit a task before starting to
create a new task. In that case the TaskComponent state will contain the
data of the editied task. Therfore the state must data must be
overriden. By using undefined to override the state variables the
defaults from the TaskDialog will be applied.