Add has_id helper function
We have several code that checks (sometimes insufficiently) if a
model/object has an id. This commits adds a clean and simple helper
function to abstract checking for an id.

In future all models should set the id to undefined if it is an empty
string. But currently doing this now it may cause several regressions.
Using the new has_id function will help to circumvent the problems.