Don't put static id into the state of the OverrideDialog
This is only a cosmetic change. The id wasn't changed therefore the old
code did work as it should. But to ensure the id isn't changed at all it
shouldn't be put into the state. Instead it must be passed as prop of
the values object to the SaveDialog. In that case id is included in the
objects passed to the onSave handler.