Fix checking for user permissions
Update EntityLink to fix checking for the current users permissions. The
permissions property is always undefined since a while. The model/entity
permissions are now parsed as user_capabilities for each model instance
and are always defined.

Don't create a link to the entity details if the capabilities have
permissions defined (areDefined is returning true) and the current
entity type can't be accessed.