Update API of Grid component
Use render props pattern for displaying Grid items.

Now it is possible to pass any kind of data to the grid items and render
the real item component from the outside via the render props pattern.
This is much more flexible and allows to adjust the props of each item
component.