Fix styling of the dashboard display title
Don't allow to wrap long titles. The complete title text should stay on
the same line. Instead ... (ellipses) should be shown if the text
overflows the possible space.

text-overflow seems to be only usable with display block and doesn't
work with display flex. Also overflow must be set to hidden.