Don't actually change port ranges before clicking "Save" in PortListDialog.
The portlist component now handles temporary lists of existing, created,
and deleted port ranges. Changes are displayed in the dialog but not
actually performed.
Temporarily created port ranges are checked, whether start<end, no or
non-numeric values are entered, and if there is an overlap with existing
and temporarily existing ranges.