Add a new Loader component
The new Loader component will replace the current data handling of the
dashboard displays. It will handle all data processing from loading the
data to passing the loaded data to the display component. The display
component will not access the redux store at the end anymore.

The Loader component can also subscribe to notifications and therefore
will be capable to reload the data on events.