Update dashboard setting data shape
Use the same object shape for the dashboard content data as expected by
the Grid layout. Create the rows and items already when the settings
have been loaded intially. This avoids unmounting and mounting of
dashboard displays when only the layout of the dashbaord displays has
changed. This commit moves the id creation for the dashboard displays
already into the action creation process.