Add reducer for saving dashboard settings
Currently the saving success and error actions are not handled because
they are not really intersting. The state is already updated before the
settings are actually saved.

If the state won't be updated in case of errors the dashboards states
would be diverge from the redux state and have to be synchonized somehow
afterwards. This would result in unexpected behavior for the user.