Fix position of the tooltip
With the performance improvement of the tooltip it is not possible to
get the ref during rendering phase anymore. Therefore the position must
be updated in componentDidUpdate which is called after render and has
all refs set. To avoid unnecessary re-render (and possible circular
rendering) don't use the state for settings the tooltip positions
instead set the positions directly via standard DOM api.

Update references to use the new React API introduced with React 16.3.