Try to circumvent circleci memory issues
Only use two jest runners. We are facing memory issues during running
the tests on circleci sometimes currently. Reading through the issues
and infos it seems to be caused by memory leaks in jest and jsdom which
should be fixed in newer versions. We are currently stuck to jest 20
because that's the version used in react-scripts 1.1.4.

Some links to details:

https://discuss.circleci.com/t/memory-problems-with-jest-and-workers/10297
https://github.com/facebook/jest/issues/3292
https://discuss.circleci.com/t/not-enough-memory-while-testing-a-node-js-appplication/12768
https://github.com/jsdom/jsdom/issues/1682