Add components for display filter selection
The FilterSelection components loads the list of filters in the store
and used the list from the store to display a filter selection in a
dialog.

The withFilterSelection HOC wraps a component into a FilterSelection if
the showFilterSelection prop is set. The HOC passes a menuEntries prop
down to the wrapped component which add a menu entry to open up the
filter selection dialog.