Fix resizing of WordCloudChart
Update to react 16.3 api (remove cWRP usage, add
getDerivedStateFromProps and use cDU for updating the displayed words).
Also ensure that the data (the word placement) only gets recalculated
if it is really necessary to avoid extra render calls.

Using d3-cloud is not optimal for react usage because it has an async
api. The code would be much easier if d3-cloud had a synchronous api
because we are only interested in the final word layout.