Use constant for task status
While doing so, also make status and ProgressBar titles translatable.

Having the getTranslatableStatus() method for tasks PLUS the exported
getTranslatableTaskStatus function was "necessary", because having it as
a Task-method is nice, but it is also needed in the StatusBar, where the
status needs translation but the task is not available (just pure status
string).