Store all loaded entities in an indexed object
When loading different lists only store the ids of the loaded entities
for the corresponding list and put all loaded entities in an byId
indexed object.

This allows to show already loaded entries (e.g. from a previous request
with a different filter).

Also this allows to get a single entity from the store.