Update dashboard settings actions and reducers
Only put default settings into the redux store when requesting loading
of the settings. Don't put the defaults into th byId part of the store.
Set object in byId part if error and success case of the loading. This
let us determine if settings have been tried to be loaded already.