Fix resizing of dashboard rows
Try to synchronize external and internal state for the current displayed
dashboard items. The synchronization should be avoided normally (and is
indeed a bad pattern in react) but I don't want to call the backend on
every height change. Therefore the external state (the redux store) is
only updated after a debounce delay of 500ms and as the result external
and internal state must be synchronized.

This change fixes the state update behavior for react 16.4 were
getDerivedStateFromProps is called not only when external props have
changed (16.3 behavior) but also after setState or forceUpdate have been
called.