Put MultiSelect into scrollable div
The TagDialog can display many resources better, although actually using
the MultiSelect does not work nicely.
I think, it makes sense to use it with small numbers of resources only
anyway. That means when the number of resources is too high to use the
DropDown efficiently, at the same time it does not make sense anymore to
change settings with it.