Use NumberField in Spinner component
Using NumberField allows to drop the state synchonization and make
Spinner a fully controlled component. State synchonization for the
displayed user input and the corresponding parsed number value is done
in NumberField now.

This commit also changes Spinner to expect a Number as value and removes
the support for a number as string.