Don't crash if dashboard display component couldn't be found
Filter dashboard items for unkown components. This shouldn't occur in
production but does occur for me because some display component names
have changed in the past and the settings still contained the old name.

This wasn't an issue some time ago but seems to crash now with an
updated version of react-beautiful-dnd.