Set current user filter as loadedFilter when loading entities
The loadedFilter is passed as filter to the children of
EntitiesContainer. When the user set a new filter in the powerfilter it
would cause a loading (and empty entities from the store). The
loadedFilter would still contain the previous filter. If the loading
takes longer the page would be re-rendered and the powerfilter would get
the old (previous) loaded filter as filter and would update the
displayed filter because it is different to the new user filter.

To avoid this behavior with long running requests the loaded filter is
set to the user filter during loading entities from backend (indicated
by undefined entitites in the props).