Don't call SaveDialog onClose if onSave was successful
Closing and saving a dialog should not be mixed. It's better if the
SaveDialog user (means parent component) differentiates between closing
and saving. It avoids race conditions where the parent dialog has
already unmounted the dialog but the close handler still has access to
some internal state of the SaveDialog.

This change requires updating the onSave handler for all SaveDialog users
to also close the dialog.