Allow to controll SaveDialog error from the parent
Update SaveDialog to React 16.4 API and drop the external error handling
by using a fully controlled error handling.

If error or onErrorClose is set on SaveDialog it is expected that the
error is controlled from the parent component. In that case the save
handler must catch the promise and pass the error to SaveDialog as a
prop.