Don't renew session in changed handler
The changed handler is not explicitly called after user interaction.
Therefore it should not renew the users session.

The current behavior in EntitesContainer is still a mix of implicit and
explicit user interaction. E.g. onFilterXyz is renewing the session but
is called only implicitly after user interaction at the powerfilter.
This may change in future by passing onInteraction down the component
tree.