Fix displaying related entities in MultiPermissionDialog
Differentiate between data that is controlled from the parent component,
data that is only displayed and data that is changed within the
MultiPermissionDialog. The related prop is controlled from the parent
component because the related entities are loaded asynchronously. id
never changes but must be put into the SaveDialog state to be included
in the data passed to the onSave handler. Same for the entityType.
entityName is constant and is only displayed. Therefore it can be
removed from the dialog state.