Change signature for loadEntities action creator
Pass gmp and filter as single arguments to the creator function. gmp now
configures the load entities and filter is the final parameter. With
this change all action creator functions should have the same signature.
Optionally the first call configures the action creator (most of the
time the gmp object is passed) and the second call passes the real
parameters.