Add a new SessionObserver component
The SessionObserver is responsible for issuing a ping request 15 seconds
after the user session has ended to return a 401 response that will
redirect to the login page. This feature is called auto logout.

Technically the SessionObserver renders a Ping component that starts a
timer on mount and removes the timer on unmount. SessionObserver uses a
super-duper reactjs feature that changing the key on a component will
always unmount and re-mount the component (see https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#recommendation-fully-uncontrolled-component-with-a-key)
for more details. This allows easy cancelling of the timer (via unmount)
instead of having to synchronize user session state changes via
componentWillReceiveProps, getDerivedStateFromProps and
componentDidUpdate like methods.