Fix creating new tag
If a tag was edited previously an id has been stored in the state.
Afterwards when the user did open the create tag dialog the id has still
been in the state. Therefore the create dialog has resulted in
overriding the previously edited tag instead of creating a new one.

This commit resets the id when opening the create tag dialog to fix this
issue.