Provide TagComponent directly for EntityTags
Don't expect handlers to be passed to EntityTags. The handlers are
currently provided by the EntityContainer component for all detail
pages. But not all detail pages have tags. For this reason and because
all detail pages are rendering the tag list table by themself its better
to provide the tag handlers directly in the EntityTags component.