Adjust entities reload interval if last load was expensive
Measure the duration of the last loading procedure and adjust the reload
timer if the duration was longer then the to be set reload interval. In
that case the new reload interval is calculated from the last duration.
The last duration is multiplied with some more or less random value to
ensure the timer takes longer then the loading procedure.