Add function to lazy translate strings
Lazy translate means it will be translated when it is converted to a
string.

Defining a module wide variable like:

const myVar = _('This should be translated')

will be translated immediately after the module is imported. Therefore
the _/translate and the i18next.t functions are called directly. If
i18next.t is not intitalized it returns undefined and as a result
undefined would my stored at myVar. Using the _l/translateLazy function
allows to postpone the call to i18next.t. It's actually called when the
object returned by _l/translateLazy is converted to a string e.g. by
calling '' + myVar or `${myVar}`.