Move rendering the label tooltips into the AllLabels component
Extract rendering the to be displayed labels of a Donut 3d chart into an
own seperate component. Code wise it's not a good decision to mix
rendering of the arcs to collect to to be displayed labels. The downside
of this change is we are iterating twice over the arcs. But now it's a
much better seperation of concerns.