Add a chartState to DataDisplay
Allow to set and get state for charts in DataDisplay. Setting the state
will also re-render the children which wouldn't be possible if a chart
keeps the state internally in it's component. At a first step the
showLegend state is put into the chart state.

In future this chart state will be saved in the redux store too.