Fix LineChart update
Only re-calculate state from width if the width has changed actually. Th
LineChart did crash because the width has been re-calculated when
hovering over the LineChart to display the tooltips. This lead to an
infinite loop.

Also fix calculation of the width. It uses the same pattern as the donut
chart now.