Fix skipping preferences in save_config_nvt_gmp
Now a generic success message is returned if no errors occur to ensure
there is a valid response.
Previously the returned message could be NULL if it was not set by
check_modify_config because there are no preferences given or the last
preference was skipped (e.g. if it's a password where the "Replace
existing ..." box was not checked).