Rework rendering of entities table rows
Render the actions components directly instead of passing them magically
via props of HOCs through all layers. With this change a Row doesn't
require to be wrapped via the withEntityRow HOC and the Actions
component is rendered in the Row directly.

Only if a table has an entity selection the actions should be wrapped
via withEntitiesActions now.