Update creating and saving a task from the TaskDialog
Add a special handler method for saving a task that checks if the task
has to be created or modified, checks if the task is
changeable/alterable, resets target, config and scanner id and closes
the dialog afterwards.

Normally and also as it was done before the correct ids are set when
OPENING the dialog already. But with loading the possible values from
the redux store and the api complexity this doesn't work anymore.