Change order of FormGroups in TargetDialog
This change lets the user see all possible credential dropdowns and 
instead "hides" the Reverse lookup Only and ReverseLookup Unify options. 
The only indication that they are there is the scrollbar on the right, 
but I don't regard this as problem as it seems to be consens that those 
settings are "advanced mode" anyway.