Improve rendering of LineChart
Don't put scales, values, min, max into state use memoization instead.

Before a re-render was triggered if data or height has changed since
last state update. This did lead first to an unnecessary render and
second to an inconsistent state for the rendering before the state
update. In this rendering phase it could be possible that the rendered
data didn't fit to the scales. This resulted in errors being logged in
the browser console.