Sort results by severity by default
For currently running scans this may be a bit cumbersome because the
results may be resorted on every reload. Opening the result details may
be difficult in this case. But nevertheless it seems to be the better
solution to sort (nearly) all tabs by the same column and show the most
important results at the top.