Add a helper function to parse a model from an element
The helper function will allow to create a model instance from a
response element with setting the entity type. This is better then
extending the fromElement static method with a entity type argument
because it should not be allowed to set an entity type when using a
specifc model instance like Task.