Add nvt columns to the cve list of a report
The cve page doesn't list each distinct cve. Instead it groups the cves
found in a nvts. Therefore is more a list of cves per nvt instead of a
list of cves. The severity also doesn't display the severity of the cves
instead it displayes the severity of the result which is indeed the
highest severity of found cves.