Use new setting for enabling/disabling the store debugging
With gmp.settings.enableStoreDebugLog the logging of the redux store can
be enabled or disabled explicitly. After setting the value to true or
value the browser needs to be reloaded to apply the store debugging.

The setting can be reset by calling gmp.settings.enableStoreDebugLog =
undefined

If the setting is undefined the store debugging is also enabled if the
log level is set to debug (gmp.settings.loglevel = 'debug').