Introduce a new component for unifying reloading
This component is responsible for initial loading, calculating the
reload timeout, reloading after timeout and reloading on request. It is
not responsible for actually loading data and passing it down to the
child components.

Most of the time passing the data is done via mapStateToProps in
conjunction with the redux store.