Update the reports details page to use Reload component
Remove all report reloading code from the ReportDetails component and
use the Reload component instead.

When initially loading a report the default results filter (either from
the user setting or the fallback default filter) is applied to get a
consistent behavior. Otherwise if the user is coming from the report
list page an unexpected filter from the data of loading the list page
would have been used. Currently this strange filter is still used but
the page will get reloaded and updated after mounting.

This unexpected behavior needs to be fixed in future to not show an
unexpected result filter at all. As a first improvement a loading
indicator should be shown.