Consider reload as inactive if the browser tab is hidden
Add handlers for visibility changed events. If the current window/tab
gets hidden the next reload interval will be "inactive" (60 seconds). If
the window/tab gets activated again to current timer will be aborted and
a new timer will be started. This new timer can be an active (3 seconds)
or default (15 seconds) reload timer depending on the current page
state.