Use loadEntityWithThreshold for loading a single report
Use the new loadEntityWithThreshold action creator in report details
page to request the report from the backend. First onyl the small report
without details is loaded. Afterwards it is checked if the report
results count is lower than the threshold from the settings and if yes
the report without details is dispatched to the redux store. Otherwise
the full report with details will be loaded.