Allow loading and storing a report with different filters
The report details page is very special. It is a combination of a
details page with a list page because the report contains a list of
results (and other data too). Therefore a report for the same id can be
very different if the results filter is changed.

This commit allows to load and store a report with a corresponding
result filter.