Add a deepFreeze function to testing utils
Copied from MDN and slightly adjusted version of a deepFreeze function.
This function can freeze an object and all of its sub-objects.
Afterwords an error is raised if somebody tries to change the object.

With deep freezing it is easier to spot mistakes when not doing a deep
copy in model parsing.