Add renew session function to useUserSessionTimeout hook
Add a new function to renew the current users session to the return
value of the useUserSessionTimeout hook. This will allow for easier
renewal of the session without having to call gmp directly.

By hiding the implementation details (calling a gmp api for a http call
that extends the current session) it allows for easier changes in future
without having to touch the using code.