Allow to renew session via hyperion
Update useUserSessionTimeout to update the session via hyperion. For now
useUserSessionTimeout also tries to extend the session in gsad if
hyperionOnly is not set. But errors doing so and the new session timeout
of gsad are ignored. Therefore it can be possible that the hyperion
session is still valid but the session in gsad not.