Fix references to reports and users in host identifiers
Fix the check for the deleted propery of the identifier which is parsed
as boolean now and got broken when swtiching to the fast xml parser. The
fast xml parser tries to be more smart by parsing integer values in
strings like '0' and '1' as integers. Therefore the comparision here
broke.