Allow to use a different name to load the page filter
The FilterProvider is used to load and determine a filter (mostly) for
list pages. In saved the current used filter at a list page in the redux
store under the gmp name e.g. task. This change allows to use a
different name for the filter saved in the redux store.