Update loading of results filter for report details
When determining the filter for loading a report and its results we need
to ensure that the user settings are considered but not to use the
stored filter of the results list page. Therefore use a different page
name in the FilterProvider to load the report results filter. The
resulting filter is passed as defaultFilter to the load function.