Try to fix permission of circleci cache
We are currently using two different images for the circleci jobs. For
building gsa and gsad we are using our own images that run under the
root user. For the js jobs we are are using images from circleci which
run under a circleci user. If the yarn.lock file is changed and the
build_gsa job is run before a js job the circle ci cache contains files
where root is the owner and with a wrong directory. Afterwards the js
jons will fail during restoring the cache because the circleci user
can't access the files from the cache.