Heavily simplify the useFormValidation hook
Remove the synchronization of state. This is always a problem and should
be avoided. The whole state must be handled outside of
useFormValidation.

Validation functions now get the new value and the state (all other
values) as a second parameter. The function must either return a string
as an error message to display a errornous validation or undefined to
indicate that the validation was successful. Alternatively an exception
may be thrown to indicate an error.

Only return on object that has three properties:
1. hasError to indicate that an error during the validation
2. errors an object of error messages
3. validate a function to check validate the form as a whole (sets
   hasError)