Fix reload interval for several pages
Using the useEntityReloadInteval or useEntitiesReloadInterval hooks
considers the active "state" of each entity model. This activate state
is only really implemented in tasks (audits) and reports. For all other
pages isActive always returns true. Using one of these hooks resulted in
a reload interval of 3 seconds which is way to often.