Don't mutate the filter in the filter provider
A filter instance should not be muated if you are not creating a new
filter or if you are not working on a copy of the original filter.
Changing an existing filter in place might have serious impact on other
places where it is used. For example if the filter is put into the redux
store the stored filter is changed too! to avoid that always use a copy
of the orignal filter before changing it.