Make simpleFilter optional in usePagination hook
Currently in all cases the simpleFilter is derived from the current
filter. Therefore there is no need to pass it to usePagination
explicitly. To provide backwards compatibility it is now optional
despite it will not be passed by a page in future.

Also add jsdoc annotations for a documentation of the hook.